🚚 move `getLocalYouwolCookie` in `backend-config.ts`
